Patch from Arne Georg Gleditsch
The latest svn revisions of filo and libpayload (r83 and r3708) conspire
to add the video console driver twice to the console_out list.  This
causes the list to become circular, with the effect that device_putchar
spins in a loop printing the first character it is given.  ("F".)

Trivial patch and .config appended.  (One might perhaps also consider
making device_putchar or console_add_* more robust in the face of this
kind of error, but I settled for this.)



git-svn-id: svn://svn.coreboot.org/filo/trunk/filo@84 205573c7-8e05-0410-a9f4-b951d841fc81