Fix EXT2_INODE_SIZE() for original format ext2 filesystems
This patch applies against r104 of filo and allows older versions of
ext2 file systems to be accessed. The problem with the current code is
that EXT2_INODE_SIZE() returns 0 for these file systems and the
EXT2_INODES_PER_BLOCK() uses that value in a calculation which causes a
div0 and sends the machine to reboot.

Below is a fix for that.

Signed-off-by: Daniel Mack <daniel@caiaq.de>
Acked-by: Peter Stuge <peter@stuge.se>


git-svn-id: svn://svn.coreboot.org/filo/trunk/filo@106 205573c7-8e05-0410-a9f4-b951d841fc81