Support Linux boot protocol v2.10
The current version of FILO is only aware of the Linux boot protocol up
to the version 2.03. This makes it impossible to, e.g., load newer
kernels compiled with CONFIG_RELOCATABLE=y because they require the boot
loader to at least support version 2.05 of the Linux boot loader
protocol. Otherwise the kernel will just panic while trying to
uncompress (because it thinks the load address is NULL).
The attached patch fixes this problem by enhancing the current
implementation to cover the newer protocol versions and respect the new
members, specifically "relocatable_kernel" and "kernel_alignment".

Signed-off-by: Mathias Krause <mathias.krause@secunet.com>
Acked-by: Peter Stuge <peter@stuge.se>


git-svn-id: svn://svn.coreboot.org/filo/trunk/filo@139 205573c7-8e05-0410-a9f4-b951d841fc81