makefile: create filo.map in separate step
Instead of copying files around and creating files make is not aware of,
use a dedicated target and rule to create filo.map. Also skip piping the
result through sort as nm has an option to sort the symbols.

Also drop the bogus dependency to libpayload as filo.elf is only the
striped version of filo.

Change-Id: If010b4ce269a47a99de06db16e0290c3fd90b559
Signed-off-by: Mathias Krause <mathias.krause@secunet.com>
Reviewed-on: http://review.coreboot.org/743
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>