makefile: build system cleanup
This patch cleans up the include hierarchies for the Makefile system to
decrease the knowledge of the main Makefile about every subdirectory.
This allows us to add new drivers without the need to patch the main
Makefile but, instead, just add a line to the appropriate Makefile.inc.

Also all standard make variables are cleaned up (e.g. CFLAGS,
CPPFLAGS,..). This, in turn, allows modifying them as needed in the
included Makefile.inc to add flags or even library dependencies.

It's no Linux Kbuild yet, but we're close ;)

Change-Id: I67be4fcdd07d6960a7e46bfb6a3c0ea41f280e2b
Signed-off-by: Mathias Krause <mathias.krause@secunet.com>
Reviewed-on: http://review.coreboot.org/746
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>