fsys_ext2: Only bail out if 64bit is really used
Smaller filesystems that have the 64-bit feature flag set should
still be compatible to the ext2 driver as long as they don't make
use of the upper 32 bits of the inode-table address. However, we
used to bail out in the exact opposite case, when the upper 32
bits are unused.

Following the paths below that check, the upper 32 bits are ignored.
Hence, the only reasonable case to continue is when they are zero.

Change-Id: I8861f30e4361f1a2c703e45f9151ecac74cd8563
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/filo/+/31242
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@google.com>