Make SATA spinup delay fully configurable
We could only toggle between a 0s and a 5s delay with CONFIG_SLOW_SATA.
On some hard disk drives spin-up takes as long as ten seconds, so make
the delay configurable. Also rename the Kconfig to SATA_SPINUP_DELAY
and update its dependencies so it can be used with any ATA driver.

Change-Id: Ia11a94479956f05206b3300cef60f06ee181579d
Signed-off-by: Paul Menzel <pmenzel@molgen.mpg.de>
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/51173
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>