Makefile: Update libpayload integration
There have been two ways to specify a libpayload location to use:

* LIBCONFIG_PATH to provide a path to libpayload sources, and
* LIBPAYLOAD_PREFIX to provide a path to a pre-compiled libpayload.

Maintaining one way usually led to breaking the other, so we try
to untangle the code parts for both methods. We keep both variables
around to hopefully stay compatible with existing build environ-
ments. Which of the methods to use is decided based on presence
of a `$(LIBPAYLOAD_PREFIX)/lib/libpayload.a`. So if both variables
are specified and LIBPAYLOAD_PREFIX is valid, it takes precedence.

If both methods fail, we try to find libpayload sources in two
standard locations:

* ../../../libpayload/
  when FILO is built by coreboot's build infrastructure, and
* ../coreboot/payloads/libpayload/
  when FILO's dir is a sibling of coreboot's (the old
  LIBCONFIG_PATH default).

When we have libpayload sources available, we use its new
`Makefile.payload`. This saves us all the dependency tracking
for libpayload and its configuration, and also much of the
compiler argument handling and make targets. This way, incre-
mental builds shouldn't be an issue even when editing libpay-
load files.

We define LIBPAYLOAD_DEFCONFIG to an empty string to override
the default path of `Makefile.payload` and disable `defconfig`
usage. This way we can keep FILO's make magic to configure
libpayload when running any `*config` target from FILO's dir.

When using a pre-compiled libpayload, we include the new
`Makefile.standalone` instead of `Makefile.payload`. This
will contain the build rules for object files and binaries.
Much like what `Makefile.payload` provides but without any
magic to handle libpayload dependencies.

Change-Id: I8bc8e37decacfa866f7f8fa33a617edda0cd97d4
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/60988
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>