drivers/sb600: Fix unused variable warning
gcc warns about a set but unused variable:

      CC      build/drivers/sb600.o
    /dev/shm/filo/drivers/sb600.c: In function ‘spi_prettyprint_status_register’:
    /dev/shm/filo/drivers/sb600.c:391:17: warning: variable ‘status’ set but not used [-Wunused-but-set-variable]
      391 |         uint8_t status;
          |                 ^~~~~~

As `msg_cdbg()` only does something if the macro `DEBUG` is defined, and
the variable `status` is only used once, only query the status register,
if the macro `DEBUG` is defined, and get rid of the variable to store
the return value. Note, this is a change in behavior.

Found-by: gcc (Debian 11.2.0-16) 11.2.0
Change-Id: I68df9f0cd1e02ede6aa92f031370cdb9b47a313c
Signed-off-by: Paul Menzel <pmenzel@molgen.mpg.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/61560
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>