loaders: Call libpayload's keyboard_disconnect() before jump
This API was added to libpayload almost a decade ago. It's generally a
good idea to return the keyboard and its controller into a known state
before starting another program. This is known to fix, for instance,
compatibility with Linux' keyboard driver when the AT translation is
in an unexpected state.

Change-Id: I5cc05b51f0d767c7f13497c06551bab29ad6c9ed
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/62234
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Thomas Heijligen <src@posteo.de>