x86/switch: Repair context backup
This swapping on the stack looks like an attempt to make it work without
the inline assembler in `context.c`. Alas, it actually wasn't disabled
as it turned out that the file doesn't go through a C preprocessor and
the `#if 0` was nothing but a deceiving comment.

Tested returning from a modified `coreinfo`.

Change-Id: I172c95522da6f89fe538059fe5d048793d35b3bc
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/62240
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>