util/kconfig: Uprev to current LTS v6.1.27
While FILO doesn't need any fancy new Kconfig features, we can benefit
from a more modern Makefile and scripts. For instance, libraries can
be detected with pkg-config now.

We follow coreboot's model of keeping the kernel's Makefile checked
in as is, and patch it at build time. This should make it easier to
uprev Kconfig in the future. Most of the files were simply copied
from the kernel tree, with these exceptions:

* `Makefile.inc` bridges our Makefile with the Kconfig one.
  It's modeled after the coreboot one, with few differences:
  $(filechk) is fully implemented (copied from the kernel's
  `scripts/Kbuild.include`) to allow incremental builds, and
  we also call respective *config targets for libpayload.

* `lexer.lex.c_shipped` and `parser.tab.[ch]_shipped` were
  generated with these commands:
  $ make CONFIG_UTIL_GENPARSER=y ${PWD}/build/util/kconfig/lexer.lex.c
  $ make CONFIG_UTIL_GENPARSER=y ${PWD}/build/util/kconfig/parser.tab.c

Change-Id: Ic8cfc7bc2877d0f14039216a5539cd104f1f0f73
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/filo/+/75093
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>